+++ NEWS.md
@@ -1,3 +1,26 @@
+# Electron.jl v2.0.0 Release Notes
+* Update Electron to version 7.1.2
+* Use artifact system
+* Drop pre Julia 1.3 support
+
+# Electron.jl v1.1.0 Release Notes
+* Add ElectronAPI
+* Fix a bug on Julia 1.3
+* Update Electron to version 4.2.12
+* Add prep_test_env function
+
+# Electron.jl v1.0.0 Release Notes
+* Update Electron to version 4.1.4
+
+# Electron.jl v0.4.0 Release Notes
+* Add load function
+
+# Electron.jl v0.3.0 Release Notes
+* Add support for callbacks from Electron to julia
+
+# Electron.jl v0.2.0 Release Notes
+* Drop julia 0.6 support, add julia 0.7 support
+
 # Electron.jl v0.1.2 Release Notes
 * Update electron version to 1.8.4
 * Make build.jl more robust